Move runtime types out of 'ktbridge' package
And into the main package.
