+++ quickjs/src/commonMain/kotlin/app/cash/quickjs/InboundCall.kt
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package app.cash.quickjs.ktbridge
+package app.cash.quickjs
 
 import kotlin.reflect.KType
 import kotlin.reflect.typeOf
