+++ quickjs/src/commonMain/kotlin/app/cash/quickjs/OutboundCall.kt
@@ -13,9 +13,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package app.cash.quickjs.ktbridge
+package app.cash.quickjs
 
-import app.cash.quickjs.ktbridge.OutboundCall.Factory
+import app.cash.quickjs.OutboundCall.Factory
 import kotlin.reflect.KType
 import kotlin.reflect.typeOf
 import okio.Buffer
